Backend/requirements.txt
Werkzeug==3.1.3
Flask==3.1.0
Jinja2==3.1.5 
MarkupSafe==3.0.2 
blinker==1.9.0 
click==8.1.8 
colorama==0.4.6 
flask==3.1.0 
itsdangerous==2.2.0 
python-dotenv==1.0.1
flask_sqlalchemy==3.1.1 
greenlet==3.1.1 
sqlalchemy==2.0.36 
typing-extensions==4.12.2
bcrypt==4.2.1
cffi==1.17.1 
cryptography==44.0.0 
jwt==1.3.1 
pycparser==2.22
PyJWT==2.10.1